QnA_Google_Palm-main/requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit
tiktoken==0.4.0
faiss-cpu==1.7.4
google-generativeai
InstructorEmbedding==1.0.1
langchain-community
torch==2.1.2
sentence-transformers==2.2.2
protobuf~=3.19.0
tqdm